requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
pandas>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
typer[all]>=0.12.0
rich>=13.7.0
streamlit>=1.36.0
